Implement as per requirements with notifications
A monitor, diagnoser and controller architecture is formed.
diagnosers are observers of monitor, controller is observer of
diagnosers. Currently implemented monitors are Mocks, which
enable unit test. A timer is also implemented as a mock.

A diagnoser and a monitor can be added without disturbing other
diagnoser and controller thus conforming to OCP.
Observer, factory, singleton, container patterns are made use of.
Liskov substitution is adhered to. SRP is followed. Cohesion is
maintained in the classes.

Signed-off-by: Harsha M M <harsha.manjulamallikarjun@in.bosch.com>